server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
ccxt==4.2.85
pycoingecko==2.3.0
python-dotenv==1.0.1
firebase-admin==6.4.0
redis==5.0.3
